Cleanup
- Move new camera code to its own place.
- Add UI toggles for the smoothing feature.
- Improve the SL voice toggle UI.
